complete success checking
